backend/utils/requirements.txt
fastapi
pydantic
PyGithub
python-dotenv
openai
